All watersheds graphed
Major copy/paste to show all stage gages graphed into their watersheds, phase 1 is grouped together in a 10 panel graph (last panel is a mix of streams that did not fit well into their respective watersheds). Email results to Caroline from IDDE gmail (to be used with a task scheduler). Do the same for Phase 2. This could probably be more efficient in some sort of function.